Fix bug with glyph lookup
An optimization was made to perform less allocations by keeping a single
global key struct. This was used for key lookup, but with replace it
will be saved into the table. Since it's a mutable global it will mess
with other keys. Therefore a fresh one should be made during replace.

Also the equal method was made 2x faster by converting to integers and
comparing. The values are already integers in floating point format to
avoid allocations. So no loss of precision should occur.