Fix Fill_stroke and Stroke_fill for the caching module
This isn't the best way to do caching, but a better way will require
larger changes, like moving the transforms to the vertex shader.

For this change the expand_fill and expand_stroke uses the current path
list to create new vertices. Because of the way caching has been
implemented we need to duplicate the paths so the path sub arrays are
not overwritten when doing two expands (stroke/fill) in a row.

In normal operation this sequence of calls is fine:

Path.begin_
rect ...
stroke
fill

because we fully resolve the vertex information by calling
Impl.stroke/fill which preserves path information in the Call
structures. For cached paths we do not immediately call Impl.stroke/fill
so the path information needs to be preserved for future draw calls.
Therefore we need to duplicate the paths so they are not overwritten be
subsequent fill/strokes.